feat: #151 Separate logging settings into logging and masking
